Refactor numbers into its own module
