+++ tests/granular/test_numbers.py
@@ -1,7 +1,8 @@
-from nlp_profiler.core import \
-    NaN, gather_whole_numbers, count_whole_numbers  # noqa
 import numpy as np
 import pytest
+
+from nlp_profiler.numbers import \
+    NaN, gather_whole_numbers, count_whole_numbers  # noqa
 
 text_with_a_number = '2833047 people live in this area'
